packages/justhtml/justhtml-1.4.0.tar.gz/justhtml-1.4.0/tests/justhtml-tests/xml_coercion.dat

#data
#xml-coercion
<!-- normal comment -->
#errors
(1,23): expected-doctype-but-got-eof

#document
| <!--  normal comment  -->
| <html>
|   <head>
|   <body>


#data
#xml-coercion
<div>normal text</div>
#errors
(1,1): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|   <body>
|     <div>
|       "normal text"


#data
#xml-coercion
<div>端ber</div>
#errors
(1,1): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|   <body>
|     <div>
|       "端ber"


#data
#xml-coercion
<div>端ber</div>
#errors
(1,1): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|   <body>
|     <div>
|       "端ber "
